update day 4

+++ backend/job-portal-backend/src/main/java/com/abhimanyu/jobportal/service/impl/UserServiceImpl.java
@@ -4,13 +4,14 @@ import com.abhimanyu.jobportal.dto.UserRequestDTO;
 import com.abhimanyu.jobportal.dto.UserResponseDTO;
 import com.abhimanyu.jobportal.entity.User;
 import com.abhimanyu.jobportal.enums.Role;
+import com.abhimanyu.jobportal.exception.UserNotFoundException;
 import com.abhimanyu.jobportal.repository.UserRepository;
 import com.abhimanyu.jobportal.service.UserService;
 
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
+
 
 @Service
 public class UserServiceImpl implements UserService {
@@ -21,7 +22,6 @@ public class UserServiceImpl implements UserService {
         this.userRepository = userRepository;
     }
 
-    // CREATE USER
     @Override
     public UserResponseDTO saveUser(UserRequestDTO dto) {
 
@@ -39,7 +39,6 @@ public class UserServiceImpl implements UserService {
         return convertToResponseDTO(savedUser);
     }
 
-    // GET ALL USERS
     @Override
     public List<UserResponseDTO> getAllUsers() {
 
@@ -49,50 +48,55 @@ public class UserServiceImpl implements UserService {
                 .toList();
     }
 
-    // GET USER BY ID
     @Override
-    public Optional<UserResponseDTO> getUserById(Long id) {
+    public UserResponseDTO getUserById(Long id) {
 
-        return userRepository.findById(id)
-                .map(this::convertToResponseDTO);
+        User user = userRepository.findById(id)
+                .orElseThrow(() ->
+                        new UserNotFoundException(
+                                "User not found with id: " + id
+                        )
+                );
+
+        return convertToResponseDTO(user);
     }
 
-    // DELETE USER
     @Override
     public void deleteUser(Long id) {
+
+        if (!userRepository.existsById(id)) {
+
+            throw new UserNotFoundException(
+                    "User not found with id: " + id
+            );
+        }
 
         userRepository.deleteById(id);
     }
 
-    // UPDATE USER
-    @Override
-    public User updateUser(Long id, User user) {
+   @Override
+public UserResponseDTO updateUser(Long id, UserRequestDTO dto) {
 
-        Optional<User> existingUserOptional =
-                userRepository.findById(id);
-
-        if (existingUserOptional.isPresent()) {
-
-            User existingUser = existingUserOptional.get();
-
-            existingUser.setFirstName(user.getFirstName());
-            existingUser.setLastName(user.getLastName());
-            existingUser.setEmail(user.getEmail());
-            existingUser.setPassword(user.getPassword());
-            existingUser.setPhone(user.getPhone());
-            existingUser.setRole(user.getRole());
-
-            return userRepository.save(existingUser);
-
-        } else {
-
-            throw new RuntimeException(
-                    "User not found with id: " + id
+    User existingUser = userRepository.findById(id)
+            .orElseThrow(() ->
+                    new UserNotFoundException(
+                            "User not found with id: " + id
+                    )
             );
-        }
-    }
 
-    // ENTITY -> RESPONSE DTO
+    existingUser.setFirstName(dto.getFirstName());
+    existingUser.setLastName(dto.getLastName());
+    existingUser.setEmail(dto.getEmail());
+    existingUser.setPassword(dto.getPassword());
+    existingUser.setPhone(dto.getPhone());
+    existingUser.setRole(Role.valueOf(dto.getRole()));
+
+    User updatedUser = userRepository.save(existingUser);
+
+    return convertToResponseDTO(updatedUser);
+}
+
+
     private UserResponseDTO convertToResponseDTO(User user) {
 
         UserResponseDTO response = new UserResponseDTO();
